Warning fixage
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@26353 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/xp/ap_Dialog_InsertTable.cpp
@@ -177,5 +177,5 @@ UT_uint32 AP_Dialog_InsertTable::getNumCols(void)
 
 float AP_Dialog_InsertTable::getColumnWidth(void)
 {
-	return UT_convertDimToInches(m_columnWidth, m_dim); 
+	return static_cast<float>(UT_convertDimToInches(m_columnWidth, m_dim));
 }
